+++ app/Jobs/Subscription/SolutionSubscriptionJob.php
@@ -1,43 +1,43 @@
 <?php
 
-namespace App\Jobs;
+namespace App\Jobs\Subscription;
 
 use App\Constants\PaymentStatus;
 use App\Contracts\PaymentGatewayContract;
-use App\Infrastructure\Persistence\Models\Payment;
+use App\Infrastructure\Persistence\Models\Subscription;
 use App\PaymentGateway\PlacetopayGateway;
 use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Foundation\Queue\Queueable;
 use Illuminate\Foundation\Bus\Dispatchable;
+use Illuminate\Foundation\Queue\Queueable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
 
-class SolutionPaymentJob implements ShouldQueue
+class SolutionSubscriptionJob implements ShouldQueue
 {
     use Dispatchable;
     use InteractsWithQueue;
     use Queueable;
     use SerializesModels;
 
-    protected Payment $payment;
+    protected Subscription $subscription;
     protected PlacetopayGateway $paymentGateway;
 
     public int $tries = 5;
-
-    public function __construct(Payment $payment)
+    public function __construct(Subscription $subscription)
     {
         $this->paymentGateway = resolve(PaymentGatewayContract::class);
-        $this->payment = $payment;
+        $this->subscription = $subscription;
     }
+
 
     public function handle(): void
     {
-        $this->paymentGateway->queryPayment($this->payment);
+        $this->paymentGateway->querySubscription($this->subscription);
     }
 
     public function failed(\Exception $exception): void
     {
-        $this->payment->status = PaymentStatus::REJECTED->value;
-        $this->payment->save();
+        $this->subscription->status = PaymentStatus::REJECTED->value;
+        $this->subscription->save();
     }
 }
